requirements.txt
# Core dependencies
numpy>=1.19.0
opencv-python>=4.5.0
matplotlib>=3.3.0
pyyaml>=5.4.0
humanize>=3.0.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.10.0
black>=20.8b1
flake8>=3.8.0
mypy>=0.800